moved project management pages outside admin folder
